requirements.txt
streamlit
beautifulsoup4
requests
langchain
openai
faiss-cpu
tiktoken